updating ft_strcmp.c

+++ ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: dleong <dleong@student.42.us.org>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/20 22:49:00 by dleong            #+#    #+#             */
-/*   Updated: 2017/09/20 23:04:26 by dleong           ###   ########.fr       */
+/*   Updated: 2017/09/24 23:27:20 by dleong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,5 +19,5 @@ int	ft_strcmp(const char *s1, const char *s2)
 	i = 0;
 	while (s1[i] != '\0' && s2[i] != '\0' && s1[i] == s2[i])
 		i++;
-	return ((unsigned char)s1[i] - (unsigned char)s2[i]);
+	return ((int)(unsigned char)s1[i] - (unsigned char)s2[i]);
 }
